% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_data}
\alias{clean_data}
\title{Clean data}
\usage{
clean_data(my_data, date, impute_bins = TRUE)
}
\arguments{
\item{my_data}{Loaded data}

\item{date}{Date to subset}

\item{impute_bins}{Inteprolate missing values}
}
\value{
data.frame
}
\description{
Cleans up the loaded data, adds a weekday and time of day column.
}
\author{
joe
}
