% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_functions.R
\name{diff_find}
\alias{diff_find}
\title{find_diff}
\usage{
diff_find(list)
}
\arguments{
\item{list}{takes a list of multiples of a single day for a single variable.
Used within the time_2_increase function and not recommeneded for use
outside of this funciton.}
}
\value{
a vector of differnces
}
\description{
find_diff
}
\author{
Joe Peskett
}
