% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_functions.R
\name{extract_features}
\alias{extract_features}
\title{extract_features}
\usage{
extract_features(list, i, d)
}
\arguments{
\item{list}{takes a list of multiples of a single day.}
}
\value{
a list of the required day and variable
}
\description{
An easy way to extract required components of the list of multiples. Not
recommended for use outside time_2_increase function.
}
\author{
Joe Peskett
}
