% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_functions.R
\name{feature_frame}
\alias{feature_frame}
\title{Feature frame}
\usage{
feature_frame(my_xts, multipliers = c(1, 1.25, 1.5, 1.75, 2))
}
\arguments{
\item{my_xts}{xts object that holds the data. Note that this can be the 
entire xts, not just a single day.}

\item{multipliers}{Numerical vector of any length. Will break if multiplier 
exceeds the maximum available in each day of data.}
}
\value{
A large list. The first level dictates the level of the multiplier, 
  second is the recording number, third is either 1 for the trough record or
  2 for the given multiplier.
}
\description{
A functional wrapper for feature_finder. Takes any number of multipliers, 
however if multiples are not present in a given day then error will be
thrown stating this.
}
\author{
Joe Peskett
}
