% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_functions.R
\name{gradients}
\alias{gradients}
\title{gradients}
\usage{
gradients(list)
}
\arguments{
\item{list}{output of time_2_increase}
}
\value{
a list for each day of the gradients
}
\description{
Returns a list of gradients from a list of differences in Bandwidth and MsM
}
\author{
Joe Peskett
}
