% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_functions.R
\name{grads}
\alias{grads}
\title{grads}
\usage{
grads(list, i)
}
\arguments{
\item{list}{a list holding a single day of differences in Bandwidth and MsM}
}
\value{
vector of gradients for that given day.
}
\description{
Helper function for gradients. May be useful elsewhere
}
\author{
Joe Peskett
}
