% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{interpolate}
\alias{interpolate}
\title{Interpolate missing values in xts time series.}
\usage{
interpolate(x, interval = "5 min", method = "linear", ...)
}
\arguments{
\item{x}{xts object}

\item{interval}{the time interval to interpolate (default = "5 min")}

\item{method}{select the interpolation method: "linear" or "merge".
default: "linear".}
}
\value{
interpolates xts object
}
\description{
Uses linear or merge interpolation.
}
\examples{

suppressWarnings(library(xts))
ts <- as.POSIXct("2018-05-31 15:00:00", tz="UTC")
x <- xts(x=1:10, order.by=seq(ts, length=10, by="5 min"))
x_broken <- x
x_broken[5] <- NA
x_fixed <- interpolate(x)
all(index(x_fixed) == index(x))
all(as.vector(x_fixed) == as.vector(x))

}
\author{
phil
}
