% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ixp.R
\docType{package}
\name{ixp-package}
\alias{ixp}
\alias{ixp-package}
\title{The ixp package holds ixp data and functions for loading,
  manipulating, analysing and visualising the data.
  
  
  
  @keywords internal}
\description{
IXP data and various helper functions
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datasciencecampus/tron}
  \item Report bugs at \url{https://github.com/datasciencecampus/tron/issues}
}

}
