% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_functions.R
\name{lag_calculator}
\alias{lag_calculator}
\title{lag_calculator}
\usage{
lag_calculator(list)
}
\arguments{
\item{list}{a list of recordings that have been selected as maximal from the
select_max() function}
}
\value{
list with differences in bandwidth and MsM
}
\description{
Function to calculate the difference in bandwidth and time between the 
windows. Be careful on what level of the list this is being run on.
}
\author{
Joe Peskett
}
