% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_and_save}
\alias{load_and_save}
\title{load csv, process and save in data/}
\usage{
load_and_save(base_dir = "", ixps = c("lon1", "lon2", "man1", "sco1",
  "nva1", "car1"))
}
\description{
convenience function.
}
\author{
phil
}
\keyword{internal}
