% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_data}
\alias{load_data}
\title{Load IXP data}
\usage{
load_data(path)
}
\arguments{
\item{path}{Location of CSV file}
}
\value{
An XTS data.frame
}
\description{
Avoids having the conversion to xts in the main notebooks.
Note this will set timezone to UTC.
}
\details{
Can set timezone of xts object with:
\code{indexTZ(man1) <- "Europe/London"}
}
\author{
joe
}
