% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{merge_interpolate}
\alias{merge_interpolate}
\title{Interpolate with merge method}
\usage{
merge_interpolate(data_sub, interval = "5 min")
}
\arguments{
\item{data_sub}{xts object}

\item{interval}{default: "5 min"}
}
\value{
Interpolated xts object
}
\description{
Alternative interpolation technique.
}
\author{
joe
}
\keyword{internal}
