% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{msm}
\alias{msm}
\title{Generate minutes since midnight}
\usage{
msm(ts, tz = "Europe/London")
}
\arguments{
\item{ts}{List of POSIXct times}
}
\value{
Integer list where each element is minute since midnight
}
\description{
Generate minutes since midnight
}
\author{
phil
}
\keyword{internal}
