% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_functions.R
\name{MsM2Date}
\alias{MsM2Date}
\title{MsM2Date}
\usage{
MsM2Date(date, MsM, shunt = -3600)
}
\arguments{
\item{date}{A date as a character in format "yyyy-mm-dd"}

\item{shunt}{A number reflecting the number of seconds to shunt the values 
by. Default is -3600 to bring it back an hour, relecting the change in 
clocks.}

\item{time}{Time in number of minutes since midnight}
}
\value{
A POSIXct object, tz = UTC, with minutes added on.
}
\description{
Quickly move from MsM to Date value - used as a utility for plotting. Not
ideal that we require the shunt argument but it's just a hack for the time
being.
}
\author{
Joe Peskett
}
