% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{my_melt}
\alias{my_melt}
\title{Melt a matrix}
\usage{
my_melt(x)
}
\arguments{
\item{x}{Matrix to melt}
}
\value{
Melted/flattened matrix
}
\description{
Who needs dplyr?
}
\examples{

x <- matrix(1:9, ncol=3)
colnames(x) <- paste0("c", 1:3)
rownames(x) <- paste0("r", 1:3)
my_melt(x)

}
\author{
phil
}
\keyword{internal}
