% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_functions.R
\name{peak_finder}
\alias{peak_finder}
\title{Peak finder}
\usage{
peak_finder(my_data, windows = list(c(15.5 * 60, 18.5 * 60), c(18.5 * 60,
  21.5 * 60)))
}
\arguments{
\item{my_data}{A list or xts object. Currently requires a time-based index 
and columns named MsM(Minutes since midnight) and Bandwidth.}

\item{windows}{in list form}
}
\value{
One long list containing the location of all the peaks and the lags
  between these and the dates for each day that has been calculated
}
\description{
This algorithm locates the peaks in a time-series. Not that at this time you
can only give one day to the function at a time - you can iterate through a 
collection of dates using a combination of sapply, do.call and rbind.
}
\author{
Joe Peskett
}
