% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{peak_plot}
\alias{peak_plot}
\title{Peak plot}
\usage{
peak_plot(my_xts, date, windows = list(c(15.5 * 60, 18.5 * 60), c(18.5 *
  60, 21.5 * 60)))
}
\arguments{
\item{my_xts}{an XTS object containing the date that you aim to plot.}

\item{date}{a single date as a character in form "yyyy-mm-dd"}

\item{windows}{Where to look for peaks, default is the same of peak_finder}
}
\value{
A ggplot of the selected date, correctly labelled with lines for the
  peaks and detailing the distance between the peaks.
}
\description{
This plots the located peaks in a time-series.
}
\author{
Joe Peskett
}
