% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{roc}
\alias{roc}
\title{Percent change from one observation to the next}
\usage{
roc(x)
}
\arguments{
\item{x}{}
}
\description{
roc_t = (Xt - Xt-1)/Xt-1
}
\examples{

roc(c(23, 10, 50, 100, 110)) # == c(NA, -0.5652174, 4, 0.1)

}
\author{
phil
}
\keyword{internal}
