% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstm_helper.R
\name{scaling_saver}
\alias{scaling_saver}
\title{scaling_saver}
\usage{
scaling_saver(data)
}
\arguments{
\item{data}{a vector for scaling}
}
\value{
a function to scale the data set, and any data sets that require the
  same values for scaling.
}
\description{
Function to scale inputs and save as objects.
}
\author{
Joe Peskett
}
