% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_functions.R
\name{select_date}
\alias{select_date}
\title{select_date}
\usage{
select_date(date = "2016-01-01")
}
\arguments{
\item{date}{This should be a character vector of a single or multiple dates 
in form "yyyy-mm-dd". The function is restricted to take lists that are
the output of feature_frame or peak_finder.}
}
\value{
A function that takes a list and returns a list.
}
\description{
Function for selecting the a date. This has been formatted to work on both 
peak_finder lists and feature_frame lists.
}
\author{
Joe Peskett
}
