% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_functions.R
\name{select_max}
\alias{select_max}
\title{select_max}
\usage{
select_max(list)
}
\arguments{
\item{list}{a list of xts objects from calling set_windows}
}
\value{
a dataframe of the required data
}
\description{
Function to collapse a window of recordings into a vector
}
\author{
Joe Peskett
}
