% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_functions.R
\name{set_windows}
\alias{set_windows}
\title{set_windows}
\usage{
set_windows(windows)
}
\arguments{
\item{windows}{a list of pairs of numbers. These should be in MsM.}
}
\value{
A function that takes an xts object and returns the defined window.
}
\description{
Function to split data into windows. Creates a closure that will subset the
data passed to it.
}
\author{
Joe Peskett
}
