% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{smooth_and_clean}
\alias{smooth_and_clean}
\title{Smooth and clean}
\usage{
smooth_and_clean(my_data, date, SMA_n = 12, impute_bins = TRUE)
}
\arguments{
\item{my_data}{Loaded data}

\item{date}{Data to subset}

\item{SMA_n}{Size of simple moving average window}

\item{impute_bins}{Interpolate missing values}
}
\value{
data.frame
}
\description{
In addition to cleaning the data, this will smooth the time-series.
}
\author{
joe
}
