% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{smooth_in_place}
\alias{smooth_in_place}
\title{Smooth xts time series in-place.}
\usage{
smooth_in_place(x, th = 0.05)
}
\arguments{
\item{x}{xts time-series}

\item{th}{Smoothing span/window. Default: 0.05 (higher values = more smooth)}
}
\value{
Smoothed xts time series
}
\description{
Uses LOESS to smooth time series in place resulting in a non-lagged series.
This can not be used as input into any predictive model since it incorporates
future values in the smoothing process.
}
\author{
phil
}
