% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{smooth_lagged}
\alias{smooth_lagged}
\title{Smooth xts time series.}
\usage{
smooth_lagged(x, th = 12)
}
\arguments{
\item{x}{xts time-series}

\item{th}{Number of periods to look back. Default: 12 (5min*12)}
}
\value{
Smoothed xts time series
}
\description{
Use EMA to create a (lagged) smooth series.
}
\author{
phi
}
