% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstm_helper.R
\name{subset_wrap}
\alias{subset_wrap}
\title{subset_wrap}
\usage{
subset_wrap(data)
}
\arguments{
\item{vector}{of univariate time series data - should be the same that was
used to create the indices that will be used in the outputted closure.}
}
\value{
a function that takes a list of indices to
}
\description{
Will use the indices created in the tstep_lag closure to create an array.
}
\author{
Joe Peskett
}
