% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_functions.R
\name{time_2_increase}
\alias{time_2_increase}
\title{time_2_increase}
\usage{
time_2_increase(list)
}
\arguments{
\item{list}{Takes}
}
\value{
A list with a space for each day. Within each day there will be two
  spaces, one for MsM and another for Bandwidth.
}
\description{
Takes the result of feature_frame and returns a list of differences in MsM 
and Bandwidth - note that the dates will be removed from this list.
}
\author{
Joe Peskett
}
