% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{ts_heatmap}
\alias{ts_heatmap}
\title{Plot bandwidth time-series as heatmap}
\usage{
ts_heatmap(ts, col.scheme = "inferno")
}
\arguments{
\item{ts}{Bandwidth time series}

\item{col.scheme}{Default: inferno}
}
\value{
ggplot
}
\description{
Alternative time-series visualisation.
}
\examples{
\dontrun{
ts_heatmap(lon1["2017"])
}

}
\author{
phil
}
