% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstm_helper.R
\name{tstep_lag}
\alias{tstep_lag}
\title{tstep_lag}
\usage{
tstep_lag(data)
}
\arguments{
\item{vector}{of univariate time series data}
}
\value{
a function that takes a time step (tstep). This function will then 
  create a list of vectors of indices that can be used to subset the 
  original input vector into the length of tstep.
}
\description{
Function to aid in the lagging of data to create x and y sets ready for
LSTM.
}
\author{
Joe Peskett
}
