% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{weekly_heatmap}
\alias{weekly_heatmap}
\title{Plot 52-week vs average-MsM heatmap}
\usage{
weekly_heatmap(ts, year, col.scheme = "inferno", daily = FALSE, border = NA)
}
\arguments{
\item{ts}{Bandwidth time series}

\item{year}{Year to subset}

\item{col.scheme}{Default: inferno}

\item{daily}{Default: FALSE. Average by day of week, else MsM.}

\item{border}{Default: NA. Colour of cell border.}
}
\value{
ggplot
}
\description{
Average MsM value for each week.
}
\examples{
\dontrun{
weekly_heatmap(lon2, year="2017", daily=FALSE)
dev.new()
weekly_heatmap(lon2, year="2017", daily=TRUE,  border="white", col.scheme="viridis")
}
}
\author{
phil
}
