% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_functions.R
\name{xts2df}
\alias{xts2df}
\title{xts2datframe}
\usage{
xts2df(xts, date)
}
\arguments{
\item{xts}{an xts object}

\item{date}{a character vector that xts can be subsetted with. Either 
"yyyy-mm-dd" for a day, "yyyy-mm-dd hh:MM" for a time, or 
"yyyy-mm-dd/yyyy-mm-dd" for a period of time. See xts subsetting rules for
more information. This can be left blank.}
}
\value{
dataframe with the subsetted data
}
\description{
Quickly take a subset by dates and returns a dataframe - used as a utilty 
for plotting
}
\author{
Joe Peskett
}
